requirement.txt
streamlit
scikit-learn
joblib
nltk
numpy
pdfplumber
python-docx

